Community 기능 구현

+++ src/main/kotlin/soma/achoom/zigg/comment/dto/CommentResponseDto.kt
@@ -5,7 +5,7 @@ import soma.achoom.zigg.user.dto.UserResponseDto
 data class CommentResponseDto(
     val commentId: Long,
     val commentMessage: String,
-    val commentLike: Long,
+    val commentLike: Int,
     val commentCreator : UserResponseDto
 ) {
 }
